==> 02_billboard_spotify.ipynb <==
# Collecting Billboard Data from Spotify
#### 11/01/2023
In this notebook, I use the Spotify API to collect songs from Billboard Year-End Top 100 playlists, and compare the results to Wikipedia data to check for correctness.

Many users have taken the time to compile public Spotify playlists according to Billboard data. Through searching online, I was able to find all the playlists and collect their links, which can be used get data through the API later:

## Finding the playlists
After about 10 minutes of searching through (and listening to) these playlists, I pasted all the links here:

https://open.spotify.com/playlist/6oMuzOETMnkHjrApJIboQl
https://open.spotify.com/playlist/1uoHvDD6ebjvmH02GtmdT9
https://open.spotify.com/playlist/741Tpr4nbbPomeeNm26zhR
https://open.spotify.com/playlist/1dyb3iCpfLYBQxdOjRuOme
https://open.spotify.com/playlist/4LDnL81lbN6ndYY86ApQAA
https://open.spotify.com/playlist/4olzlochc7bKxI9meYVBQ5
https://open.spotify.com/playlist/0NT0tq924n07BCDt9f5ZbL
https://open.spotify.com/playlist/0D8RVLZQJKbt1txw7fXCoA
https://open.spotify.com/playlist/0S6g0tqZNT7xSdlNQzU2OQ
https://open.spotify.com/playlist/7FnOEoPb0biPkxYYAoXO8Q
https://open.spotify.com/playlist/4aUY170nZ3mhkzMpTAXDv2
https://open.spotify.com/playlist/2z3eLip2NlV9quzTEm37cW
https://open.spotify.com/playlist/0W42jv9ZJTwCL0cLuLbTfX
https://open.spotify.com/playlist/1KK0RvFmgsUkZ8zELRZgjS
https://open.spotify.com/playlist/2trgZsxRpWX7sq28yHC40u
https://open.spotify.com/playlist/6LYxiUgw87zsDPqU0sdalZ
https://open.spotify.com/playlist/3JbWD8OGutoTKUbR3RvR8u
https://open.spotify.com/playlist/2XPEN88QyrPQ9zGqS8uS2x
https://open.spotify.com/playlist/4MCwbKJPbmfKY08aTyur9e
https://open.spotify.com/playlist/6sxay2eKIKLxHCuytNSohv
https://open.spotify.com/playlist/2jDNKyd7Fs8Zf3pLVkCasY
https://open.spotify.com/playlist/1bjAgktyYGqKJXynU4H8SO
https://open.spotify.com/playlist/6tsGGO4rewMHMjbvSUHHc7

links = str.split("https://open.spotify.com/playlist/6oMuzOETMnkHjrApJIboQl https://open.spotify.com/playlist/1uoHvDD6ebjvmH02GtmdT9 https://open.spotify.com/playlist/741Tpr4nbbPomeeNm26zhR https://open.spotify.com/playlist/1dyb3iCpfLYBQxdOjRuOme https://open.spotify.com/playlist/4LDnL81lbN6ndYY86ApQAA https://open.spotify.com/playlist/4olzlochc7bKxI9meYVBQ5 https://open.spotify.com/playlist/0NT0tq924n07BCDt9f5ZbL https://open.spotify.com/playlist/0D8RVLZQJKbt1txw7fXCoA https://open.spotify.com/playlist/0S6g0tqZNT7xSdlNQzU2OQ https://open.spotify.com/playlist/7FnOEoPb0biPkxYYAoXO8Q https://open.spotify.com/playlist/4aUY170nZ3mhkzMpTAXDv2 https://open.spotify.com/playlist/2z3eLip2NlV9quzTEm37cW https://open.spotify.com/playlist/0W42jv9ZJTwCL0cLuLbTfX https://open.spotify.com/playlist/1KK0RvFmgsUkZ8zELRZgjS https://open.spotify.com/playlist/2trgZsxRpWX7sq28yHC40u https://open.spotify.com/playlist/6LYxiUgw87zsDPqU0sdalZ https://open.spotify.com/playlist/3JbWD8OGutoTKUbR3RvR8u https://open.spotify.com/playlist/2XPEN88QyrPQ9zGqS8uS2x https://open.spotify.com/playlist/4MCwbKJPbmfKY08aTyur9e https://open.spotify.com/playlist/6sxay2eKIKLxHCuytNSohv https://open.spotify.com/playlist/2jDNKyd7Fs8Zf3pLVkCasY https://open.spotify.com/playlist/1bjAgktyYGqKJXynU4H8SO https://open.spotify.com/playlist/6tsGGO4rewMHMjbvSUHHc7")
len(links)

In any given link, the playlist ID is the text after ".../playlist/..."

# Extract the playlist IDs
playlist_ids = [playlist_id.replace("https://open.spotify.com/playlist/", '') for playlist_id in links]
print(playlist_ids)

Now that we have all these IDs, we can start our Spotify searches:

## Gathering playlist data

import pandas as pd
from tqdm import tqdm
import requests
import base64
import re
import json

The first thing to do is authenticate myself and get an access token, so that I can complete searches (I'll have to run this code every hour or so):

# Request setup
CLIENT_ID = '526daa7bf54e45ee891e83cc0f806f10'
CLIENT_SECRET = '25f3c9ec9d21478083879659c40455b5'
AUTH_URL = 'https://accounts.spotify.com/api/token'
headers = {
    'Authorization': 'Basic ' + base64.b64encode(f'{CLIENT_ID}:{CLIENT_SECRET}'.encode()).decode()
}
data = {
    'grant_type': 'client_credentials'
}
auth_response = requests.post(AUTH_URL, headers=headers, data=data)

# Get the access token
access_token = auth_response.json().get('access_token')

# Set up headers for future API calls
headers = {
    'Authorization': 'Bearer ' + access_token
}

# Check if we successfully fetched the token
if access_token:
    print("Successfully authenticated!")
else:
    print("Failed to authenticate")

Now, we can begine the search.

def get_playlist_data(playlist_id, token):
    """
    Function to get the playlist data (a json file) given a playlist ID.
    """
    BASE_URL = "https://api.spotify.com/v1/playlists/"
    HEADERS = {
        "Authorization": f"Bearer {token}",
        "Content-Type": "application/json"
    }

    URL = BASE_URL + f"{playlist_id}/tracks"
    response = requests.get(URL, headers=HEADERS)
    return response.json()


# Apply the above function for all playlists
all_playlist_data = [get_playlist_data(playlist_id, access_token) for playlist_id in tqdm(playlist_ids)]

Now, that we have all the playlist data, let's see if it matches up with our ground truth.

## Finding mismatched songs
The first thing to check is whether or not the Spotify data has the right number of songs in each year:

for i in range(23):
    print(2000 + i, len(all_playlist_data[i]['items']), end='     ')

It doesn't... 2001 and 2012 are both missing one song. I'll keep that in mind.

I'm not going to get into exactly how I verified each song, as it was a bit mind-numbing. But all the code is below:

wiki_songs = pd.read_csv('wiki_songs.csv')
wiki_spotify_songs = wiki_songs[['Title']]

spotify_songs = []
for playlist in all_playlist_data:
    items = playlist['items']
    for i in items:
        spotify_songs.append(i['track']['name'])

wiki_spotify_songs['Spotify'] = pd.Series(spotify_songs)

def simplify_title(title):
    if type(title) != str:
        return title
    title = title.lower()
    title = title = re.sub(r'[^a-z0-9\s]', '', title) # removes any special characters
    title = re.sub(r'\(.*?\)', '', title)  # removes content within ()
    title = re.sub(r'\[.*?\]', '', title)  # removes content within []
    title = title.strip()
    return title[:6] if len(title) >= 5 else title

wiki_spotify_songs['Title'] = wiki_spotify_songs['Title'].apply(simplify_title)
wiki_spotify_songs['Spotify'] = wiki_spotify_songs['Spotify'].apply(simplify_title)

#wiki_spotify_songs[wiki_spotify_songs['Title'] != wiki_spotify_songs['Spotify']]
#wiki_spotify_songs['Spotify'] = wiki_spotify_songs['Spotify'].shift(1)
#wiki_spotify_songs[1281:][wiki_spotify_songs['Title'] != wiki_spotify_songs['Spotify']]

And after searching through the long list, it turns out the only two edits I need to make are these two songs:

_Get Over Yourself_ by Eden's Crush, ranking 92 in 2001 (The official song isn't on Spotify)

_Birthday Cake_ by Rihanna, ranking 79 in 2012 (this song is really just not that good, maybe that's why it wasn't included :P)

## Inserting the missing data

The best way to retrieve the song data from Spotify is by making a query based on track name and artist name:

def get_track_data(track_name, artist_name, token):
    headers = {
        "Authorization": f"Bearer {token}"
    }
    
    query = f"track:{track_name} artist:{artist_name}"
    params = {
        "q": query,
        "type": "track",
        "limit": 1
    }
    
    response = requests.get("https://api.spotify.com/v1/search", headers=headers, params=params)
    
    # If we find the song...
    if response.status_code == 200:
        return response.json()
        if data['tracks']['items']:
            return data['tracks']['items'][0]
    else:
        print(f"Error {response.status_code}: {response.text}")
        return None

get_over_yourself_data = get_track_data("Get Over Yourself", "BFM Hits", access_token)
get_over_yourself_data = {'track': get_over_yourself_data['tracks']['items'][0] }

birthday_cake_data = get_track_data("Birthday Cake", "Rihanna", access_token)
birthday_cake_data = {'track': birthday_cake_data['tracks']['items'][0] }

It took a lot of trial and error to get the song data to be the same format as the rest of the songs in the playlists, but eventually, I got it to match and was ready to insert:

all_playlist_data[1]['items'].insert(91, get_over_yourself_data)
all_playlist_data[12]['items'].insert(78, birthday_cake_data)

# Finally, let's save this to a json file for later use
with open('all_playlist_data.json', 'w', encoding='utf-8') as f:
    json.dump(all_playlist_data, f, ensure_ascii=False, indent=4)

## Next steps

Glad this part is done! Now, it's time to make our datasets.
