==> 03_billboard_basic.ipynb <==
# Preparing the "Basic" Billboard Dataset
#### 11/01/23
In this dataset, I use the previously made JSON file to create the basic dataset.

## Reading in the data

import pandas as pd
import json

def load_from_json(file_name):
    with open('all_playlist_data.json', 'r', encoding='utf-8') as f:
        return json.load(f)

all_playlist_data = load_from_json('all_playlist_data.json')

def get_track_info(track_data):
    track_id = track_data['id']
    title = track_data['name']
    artist = track_data['artists'][0]['name']
    album = track_data['album']['name']
    release_date = track_data['album']['release_date']
    popularity = track_data['popularity']
    duration_ms = track_data['duration_ms']
    explicit = track_data['explicit']
    
    return track_id, title, artist, album, release_date, popularity, duration_ms, explicit

def get_playlist_info(playlist_items):
    return [get_track_info(playlist_items[i]['track']) for i in range(len(playlist_items))]

all_songs_data = []
for i in range(23):
    all_songs_data += get_playlist_info(all_playlist_data[i]['items'])

df = pd.DataFrame(all_songs_data)
df.head()

## Pre-processing
All we have to do now is name the columns.

df.columns = ['track_id', 'title', 'artist', 'album', 'release_date', 'popularity', 'duration_ms', 'explicit']
df.head()

df['year'] = df.index // 100 + 2000
df['ranking'] = df.index % 100 + 1
df

df.to_csv('billboard_basic.csv', index=False)

## Next steps

Now we have the basic dataset. Next, we'll make the advanced dataset, which will include audio features.
