==> 04_billboard_advanced.ipynb <==
# Creating the "Advanced" Dataset
#### 11/02/2023
In this notebook, I again use the Spotify API to get the audio feature data from each song. After collecting this data, I combine it with the basic dataset to get the advanced dataset, my final task for this project.

## Reading in the track IDs
In order to collect the audio features for each song, we'll need to get the track IDs. Let's pull them from the basic dataset we made last time:

import pandas as pd
import requests
import json
from tqdm import tqdm
import time

df = pd.read_csv('billboard_basic.csv')
track_ids = df['track_id']

## Gathering all the features
Using these IDs, we can request Spotify for the audio features. Firstly, let's do the authentication:

### Authentication

import base64

# Request setup
CLIENT_ID = '526daa7bf54e45ee891e83cc0f806f10'
CLIENT_SECRET = '25f3c9ec9d21478083879659c40455b5'
AUTH_URL = 'https://accounts.spotify.com/api/token'
headers = {
    'Authorization': 'Basic ' + base64.b64encode(f'{CLIENT_ID}:{CLIENT_SECRET}'.encode()).decode()
}
data = {
    'grant_type': 'client_credentials'
}
auth_response = requests.post(AUTH_URL, headers=headers, data=data)

# Get the access token
access_token = auth_response.json().get('access_token')

# Set up headers for future API calls
headers = {
    'Authorization': 'Bearer ' + access_token
}

# Check if we successfully fetched the token
if access_token:
    print("Successfully authenticated!")
else:
    print("Failed to authenticate")


Now let's define a function to collect the audio features based on a given index, and save these audio features to a json file:

BASE_URL = "https://api.spotify.com/v1/"

def save_audio_features(index):
    try:
        with open("cache.json", "r") as cache_file:
            cache = json.load(cache_file)
    except (FileNotFoundError, json.JSONDecodeError):
        # If the cache file doesn't exist or is empty, initialize an empty cache
        cache = {}
    
    # If the track_id is found in cache, skip this index
    if str(index) in cache:
        return
    
    URL = BASE_URL + f"audio-features/{track_ids[index]}"
    response = requests.get(URL, headers=headers)
    features = response.json()
    
    if 'error' in features:
        raise Exception("Too many requests")
    
    cache[index] = features
    with open("cache.json", "w") as cache_file:
        json.dump(cache, cache_file)

Now here's the hard part. Since I run the risk of getting a 429 error and killing the kernel, I'll break it up into parts and do this over the course of a few hours.

for index in tqdm(range(2000, 2300)):
    save_audio_features(index)

# Let's make a new dataframe from the json file
df2 = pd.read_json('cache.json', orient='index')

df2

Finally, we've gotten all the song features. Now, it's time to connect them to their respective songs.

## Concatenating the two dataframes
This part should be easy since the two dataframes have matching indicies:

combined_df = pd.concat([df, df2], axis=1)
combined_df

## Pre-processing (final touches)
Lastly, let's remove some columns that we don't need.

billboard_advanced = combined_df.drop(str.split('type id uri track_href analysis_url'), axis=1)

billboard_advanced.to_csv('billboard_advanced.csv', index=False)
